Move openshot_rc.py to classes/
